backend_archive_20260213/requirements.txt

# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# HTTP & Async
httpx>=0.26.0
aiohttp>=3.9.0
playwright>=1.41.0
openai>=1.0.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# URL & HTML Parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# SSL & Security
pyOpenSSL>=24.0.0
cryptography>=41.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
sqlmodel>=0.0.14
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Authentication & Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0
validators>=0.22.0

# Monitoring & Email
apscheduler>=3.10.0
fastapi-mail>=1.4.0

# Image Processing
Pillow>=10.0.0
numpy>=1.26.0

# Redis/Celery (Required for Docker)
redis>=5.0.1
celery>=5.3.6

# Payments
stripe>=7.0.0
